refactor: dedup landing ContentRail and CLI gateway-url resolution
Consolidate ContentRail + CompactContentRail into a single component with
'page' | 'inset' | 'compact' variants, and delete CompactContentRail.

Move resolveGatewayUrl (plus GATEWAY_DEFAULT_URL constant) into cli-core
so chat.ts and eval.ts share one implementation. Delete the pure re-export
shims in packages/cli/src/api/ and import directly from @lobu/cli-core.
