--- packages/cli/src/api/context.ts
@@ -1,12 +1,0 @@
-export {
-  addContext,
-  DEFAULT_CONTEXT_NAME,
-  getCurrentContextName,
-  type LobuContextConfig,
-  type LobuContextEntry,
-  LOBU_CONFIG_DIR,
-  loadContextConfig,
-  type ResolvedContext,
-  resolveContext,
-  setCurrentContext,
-} from "@lobu/cli-core";
